+++ packages/core/src/types/index.ts
@@ -529,3 +529,32 @@ export interface IChain {
    */
   write(call: unknown): Promise<any>;
 }
+
+export enum HandlerRole {
+  INPUT = "input",
+  OUTPUT = "output",
+  ACTION = "action",
+}
+
+/**
+ * A single interface for all Inputs, Outputs.
+ */
+export interface IOHandler {
+  /** Unique name for this handler */
+  name: string;
+
+  /** "input" | "output" | (optionally "action") if you want more roles */
+  role: HandlerRole;
+
+  /** For input handlers with recurring scheduling */
+  interval?: number;
+
+  /** The schema for the input handler */
+  schema: z.ZodType<any>;
+
+  /** Next run time (timestamp in ms); for input scheduling. */
+  nextRun?: number;
+
+  /** The main function. For inputs, no payload is typically passed. For outputs, pass the data. */
+  handler: (payload?: unknown) => Promise<unknown>;
+}
